pyproject.toml
[tool.poetry]
name = "landmark"
version = "0.1.0"
description = ""
authors = ["jayhong99 <jayhong7200@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.12.1"
torchvision = "^0.13.1"
numpy = "^1.23.1"
pandas = "^1.4.3"
PyYAML = "^6.0"
joblib = "^1.1.0"
Pillow = "^9.2.0"
scikit-learn = "^1.1.2"

[tool.poetry.dev-dependencies]
ipykernel = "^6.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
